Add cast to silence warning on MSVC
Address should be 5 bytes before the end of the `CALL` instruction,
expressed as a relative offset, with unsigned wraparound semantics.

+++ test/OP2Memory.test.cpp
@@ -43,7 +43,7 @@ TEST_F(Op2RelinkCallTest, GoodOpcodeSelfCall) {
   // Opcode should stay the same
   EXPECT_EQ(CallOpcode, callInstruction.opcode);
   // Relative address should point back to itself
-  EXPECT_EQ(-sizeof(CallInstruction), callInstruction.relativeAddress);
+  EXPECT_EQ(static_cast<uint32_t>(-sizeof(CallInstruction)), callInstruction.relativeAddress);
 }
 
 TEST_F(Op2RelinkCallTest, BadOpcode) {
